refactor(routes): consolidate private routes into user routes module
Move private routes from separate plugin to user routes module to simplify routing structure and reduce duplication. Update default port in env example to 4000.
