+++ server/.env.example
@@ -1,7 +1,7 @@
 # Application Configuration
 NODE_ENV=development
-PORT=3000
-API_BASE_URL=http://localhost:3000
+PORT=4000
+API_BASE_URL=http://localhost:4000
 LOG_LEVEL=info
 
 # Database Configuration
