+++ server/src/app/routes/index.ts
@@ -2,4 +2,3 @@ export * from './types';
 export * from './registry';
 export * from './manager';
 export { publicRoutesPlugin } from './groups/public';
-export { privateRoutesPlugin } from './groups/private';
